update on task 14

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	int i, j, k, next;
+	long int i, j, k, next;
 
 	j = 1;
 	k = 2;
